requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.1
requests==2.32.3
openai==1.68.2
PyYAML==6.0.2
